fix #19373 move all hairpins; insert measure always acts in master score
